Added Entity, Butterfly, and Entity Manager Class
Finished making the visible butterfly

Fixed an issue where tileNums was pointing to initTileNums so picked flowers wouldn't turn into grass

* Still needs to interact with flower, pushback player, flee, swat, etc.

Inheritance is set up well for all entities

* Continue to work on maintaining proper inheritance

* Make sure you don't cause cyclic dependency issues with the EntityManager class

UP NEXT: Continue to work on Butterfly mechanics, maybe starting with flower interaction.

My idea is to make the butterfly stop at the up position on top of the flower and after a couple seconds picks it and continues on its path.

The player should not be able to interact with a flower that an entity is interacting with, make sure pickable is set to false.

Adversely, the entities should be able to keep pursuing a flower the player is picking.